+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/mdp/__init__.py
@@ -5,5 +5,6 @@ Copyright (c) 2024, Felipe Mohr Santos
 
 from omni.isaac.lab_tasks.manager_based.locomotion.velocity.mdp import *  # noqa: F401, F403
 
+from .curriculums import *  # noqa: F401, F403
 from .observations import *  # noqa: F401, F403
 from .rewards import *  # noqa: F401, F403