+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/mdp/curriculums.py
@@ -1,0 +1,74 @@
+"""
+License: MIT License
+Copyright (c) 2024, Felipe Mohr Santos
+"""
+
+from __future__ import annotations
+
+import torch
+from collections.abc import Sequence
+from typing import TYPE_CHECKING, Any
+
+if TYPE_CHECKING:
+    from omni.isaac.lab.envs import ManagerBasedRLEnv
+    from omni.isaac.lab.managers import SceneEntityCfg
+
+
+def modify_event_parameter(
+    env: ManagerBasedRLEnv,
+    env_ids: Sequence[int],
+    term_name: str,
+    param_name: str,
+    value: Any | SceneEntityCfg,
+    num_steps: int,
+) -> torch.Tensor:
+    """Curriculum that modifies a parameter of an event at a given number of steps.
+
+    Args:
+        env: The learning environment.
+        env_ids: Not used since all environments are affected.
+        term_name: The name of the event term.
+        param_name: The name of the event term parameter.
+        value: The new value for the event term parameter.
+        num_steps: The number of steps after which the change should be applied.
+
+    Returns:
+        torch.Tensor: Whether the parameter has already been modified or not.
+    """
+    if env.common_step_counter > num_steps:
+        # obtain term settings
+        term_cfg = env.event_manager.get_term_cfg(term_name)
+        # update term settings
+        term_cfg.params[param_name] = value
+        env.event_manager.set_term_cfg(term_name, term_cfg)
+        return torch.ones(1)
+    return torch.zeros(1)
+
+
+def disable_termination(
+    env: ManagerBasedRLEnv,
+    env_ids: Sequence[int],
+    term_name: str,
+    num_steps: int,
+) -> torch.Tensor:
+    """Curriculum that modifies the push velocity range at a given number of steps.
+
+    Args:
+        env: The learning environment.
+        env_ids: Not used since all environments are affected.
+        term_name: The name of the termination term.
+        num_steps: The number of steps after which the change should be applied.
+
+    Returns:
+        torch.Tensor: Whether the parameter has already been modified or not.
+    """
+    env.command_manager.num_envs
+    if env.common_step_counter > num_steps:
+        # obtain term settings
+        term_cfg = env.termination_manager.get_term_cfg(term_name)
+        # Remove term settings
+        term_cfg.params = dict()
+        term_cfg.func = lambda env: torch.zeros(env.num_envs, device=env.device, dtype=torch.bool)
+        env.termination_manager.set_term_cfg(term_name, term_cfg)
+        return torch.ones(1)
+    return torch.zeros(1)